+++ R/download.R
@@ -158,8 +158,9 @@ nc_data_dl <- function(collections = NULL, project_ids = NULL,
     } else no_access <- c()
 
     if(length(no_access) == 0) {
-      stop("These collections have no data that match these filters",
-           call. = FALSE)
+      warning("These collections have no data that match these filters",
+              call. = FALSE)
+      return(invisible())
     } else {
       stop("You do not have permission to access these collections (",
            paste0(no_access, collapse = ", "), ")", call. = FALSE)