+++ data-raw/data_creation.R
@@ -77,12 +77,12 @@ queries <- dplyr::tribble(
 
 # Testing Data ------------------------------------------------------------
 test_rc <- nc_data_dl(request_id = 152518, fields_set = "core",
-                      username = "sample", info = "sample_data") %>%
+                      username = "sample", info = "sample_data") |>
   dplyr::filter(CommonName %in% c("Monarch",
                                   "Black Swallowtail",
                                   "Red Admiral"),
-                AllSpeciesReported == "Yes") %>%
-  dplyr::mutate(presence = as.numeric(ObservationCount > 0)) %>%
+                AllSpeciesReported == "Yes") |>
+  dplyr::mutate(presence = as.numeric(ObservationCount > 0)) |>
   format_dates()
 
 # Field order - Non BMDE fields
@@ -108,9 +108,10 @@ usethis::use_data(bcch, internal = FALSE, overwrite = TRUE)
 hofi <- nc_data_dl(species = 20350, username = "sample", info = "pkg_data")
 usethis::use_data(hofi, internal = FALSE, overwrite = TRUE)
 
-
+unlink(file.path("inst", "extdata", "bcch.nc"))
 nc_data_dl(request_id = 152543, username = "sample", 
            sql_db = file.path("inst", "extdata", "bcch"))
 
+unlink(file.path("inst", "extdata", "hofi.nc"))
 nc_data_dl(species = 20350, username = "sample", info = "pkg_data",
            sql_db = file.path("inst", "extdata", "hofi"))